extend data return object
